pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "orion-finance-sdk"
version = "0.3.0"
requires-python = ">=3.11"
readme = "README.md"
authors = [
    { name = "Orion Finance" },
]

dependencies = [
    "python-dotenv>=1.1.0,<2.0.0",
    "numpy>=2.2.0,<3.0.0",
    "typer>=0.16.0,<1.0.0",
    "web3>=7.12.0,<8.0.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.1.0,<4.2.0",
    "ruff>=0.9.10, < 1.0.0",
    "pydocstyle>=6.3.0,<6.4.0",
    "pytest>=8.3.5,<8.4.0",
    "pydeps>=3.0.0,<4.0.0"
]

[project.scripts]
orion = "orion_finance_sdk.__main__:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
